+++ ext/hmsSerializer.hpp
@@ -150,16 +150,16 @@ namespace ext
         bool assignOp<bool>(const Json::Value* pSource, bool& pDestination);
         
         template <>
-        bool assignOp<int32_t>(const Json::Value* pSource, int& pDestination);
+        bool assignOp<int32_t>(const Json::Value* pSource, int32_t& pDestination);
         
         template <>
-        bool assignOp<int64_t>(const Json::Value* pSource, long long int& pDestination);
+        bool assignOp<int64_t>(const Json::Value* pSource, int64_t& pDestination);
         
         template <>
-        bool assignOp<uint32_t>(const Json::Value* pSource, unsigned& pDestination);
+        bool assignOp<uint32_t>(const Json::Value* pSource, uint32_t& pDestination);
         
         template <>
-        bool assignOp<uint64_t>(const Json::Value* pSource, unsigned long long int& pDestination);
+        bool assignOp<uint64_t>(const Json::Value* pSource, uint64_t& pDestination);
         
         template <>
         bool assignOp<float>(const Json::Value* pSource, float& pDestination);
